Refactor commands out of the component
